vad/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.1
boto3==1.26.141

pyannote.audio==2.1.1
pydub==0.25.1
requests==2.31.0